stack relocated to checker
